% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBR-methods.R
\docType{methods}
\name{mbrBootstrap,MBR-method}
\alias{mbrBootstrap,MBR-method}
\alias{mbrBootstrap}
\title{Inference of consensus transcriptional networks.}
\usage{
\S4method{mbrBootstrap}{MBR}(object, verbose = TRUE, ...)
}
\arguments{
\item{object}{A processed objec of class \linkS4class{MBR} evaluated by the 
method \code{\link{mbrPermutation}}.}

\item{verbose}{A single logical value specifying to display detailed 
messages (when verbose=TRUE) or not (when verbose=FALSE).}

\item{...}{Additional arguments passed to the 
\code{\link{tni.bootstrap}} function.}
}
\value{
An \linkS4class{MBR} object with two consensus mutual information 
matrices, one in each "TNI" slot.
}
\description{
This function takes an MBR object and computes two consensus transcriptional 
networks.
}
\examples{
##--- load a dataset for demonstration
data("dt4rtn", package = "RTN")
gexp <- dt4rtn$gexp
annot <- dt4rtn$gexpIDs
tfs <- dt4rtn$tfs[c("IRF8","IRF1","PRDM1","AFF3","E2F3")]

##--- run mbrPreprocess
rmbr <- mbrPreprocess(gexp=gexp, regulatoryElements = tfs, 
rowAnnotation=annot)

##--- run mbrPermutation (nPermutations>=1000)
rmbr <- mbrPermutation(rmbr, nPermutations=10)

##--- run mbrBootstrap (nPermutations>=100)
rmbr <- mbrBootstrap(rmbr, nBootstrap=10)

}
