% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBR-methods.R
\docType{methods}
\name{mbrDpiFilter,MBR-method}
\alias{mbrDpiFilter,MBR-method}
\alias{mbrDpiFilter}
\title{A filter based on the Data Processing Inequality (DPI) algorithm.}
\usage{
\S4method{mbrDpiFilter}{MBR}(object, verbose = TRUE, ...)
}
\arguments{
\item{object}{A processed object of class \linkS4class{MBR} evaluated by 
the methods
 \code{\link{mbrPermutation}} and 
 \code{\link{mbrBootstrap}}.}

\item{verbose}{A single logical value specifying to display detailed 
messages (when verbose=TRUE) or not (when verbose=FALSE).}

\item{...}{Additional arguments passed to the 
\code{\link{tni.dpi.filter}} function.}
}
\value{
An \linkS4class{MBR} object with two DPI-filtered mutual information 
matrices, one in each "TNI" slot.
}
\description{
This function takes an MBR object and computes two transcriptional networks 
filtered by the data processing inequality algorithm.
}
\examples{
##--- load a dataset for demonstration
data("dt4rtn", package = "RTN")
gexp <- dt4rtn$gexp
annot <- dt4rtn$gexpIDs
tfs <- dt4rtn$tfs[c("IRF8","IRF1","PRDM1","AFF3","E2F3")]

##--- run mbrPreprocess
rmbr <- mbrPreprocess(gexp=gexp, regulatoryElements = tfs, 
rowAnnotation=annot)

##--- run mbrPermutation (set nPermutations>=1000)
rmbr <- mbrPermutation(rmbr, nPermutations=10)

##--- run mbrBootstrap (set nBootstrap>=100)
rmbr <- mbrBootstrap(rmbr, nBootstrap=10)

##--- run mbrDpiFilter
rmbr <- mbrDpiFilter(rmbr)

}
