% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBR-methods.R
\docType{methods}
\name{mbrDuals,MBR-method}
\alias{mbrDuals,MBR-method}
\alias{mbrDuals}
\title{A summary for results from the MBR methods.}
\usage{
\S4method{mbrDuals}{MBR}(object, supplementaryTable = NULL,
  evidenceColname = NULL, verbose = TRUE)
}
\arguments{
\item{object}{A processed object of class \linkS4class{MBR} evaluated by the 
method \code{\link[RTNduals:mbrAssociation]{mbrAssociation}}.}

\item{supplementaryTable}{An optional 'data.frame' with three columns 
representing (1) regulatory elements of 'TNI1', (2) regulatory elements 
of 'TNI2', and (3) external evidences between the regulatory elements.}

\item{evidenceColname}{A single character value specifying a column in 
the 'supplementaryTable'.}

\item{verbose}{A single logical value specifying to display detailed 
messages (when verbose=TRUE) or not (when verbose=FALSE).}
}
\value{
An \linkS4class{MBR} object with an updated 'data.frame' in the slot 
'results' listing the input additional evidences.
}
\description{
This function lists the inferred 'dual regulons' and, if available, 
adds external evidences.
}
\examples{
##--- load a dataset for demonstration
data("dt4rtn", package = "RTN")
gexp <- dt4rtn$gexp
annot <- dt4rtn$gexpIDs
tfs1 <- dt4rtn$tfs[c("IRF8","IRF1","PRDM1","AFF3","E2F3")]
tfs2 <- dt4rtn$tfs[c("HCLS1","STAT4","STAT1","LMO4","ZNF552")]

##--- run mbrAssociation
rmbr <- mbrPreprocess(gexp=gexp, regulatoryElements1 = tfs1, 
regulatoryElements2=tfs2, rowAnnotation=annot)
rmbr <- mbrPermutation(rmbr, nPermutations=10)
rmbr <- mbrBootstrap(rmbr, nBootstrap=10)
rmbr <- mbrDpiFilter(rmbr)
rmbr <- mbrAssociation(rmbr, minRegulonSize = 15, prob=0.75)
rmbr <- mbrDuals(rmbr)

##--- check results
results <- mbrGet(rmbr, what="dualsInformation")

##--- add supplementary evidences
## here we build a 'toy' example using the 'rnorm' function for 
## demonstration purposes only!
supplementaryTable <- results[,c("Regulon1","Regulon2")]
supplementaryTable$ToyEvidence <- rnorm(nrow(results))
supplementaryTable

##--- add supplementary evidences with brDuals function
rmbr <- mbrDuals(rmbr, supplementaryTable=supplementaryTable,
evidenceColname = "ToyEvidence")

##--- check updated results
mbrGet(rmbr, what="dualsInformation")

}
