% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBR-methods.R
\docType{methods}
\name{mbrGet,MBR-method}
\alias{mbrGet,MBR-method}
\alias{mbrGet}
\title{Get information from individual slots in MBR object.}
\usage{
\S4method{mbrGet}{MBR}(object, what = "status")
}
\arguments{
\item{object}{A preprocessed object of class \linkS4class{MBR}}

\item{what}{a single character value specifying which information should be 
retrieved from the slots. Options: "TNI", "regulatoryElements", 
"dualRegulons", "results", "para", "summary", 
"status", "dualsCorrelation", "dualsOverlap", and "dualsCorMatrix"}
}
\value{
Content from slots in the \linkS4class{MBR} object
}
\description{
Get information from individual slots in an MBR object and any available 
results from previous analysis.
}
\examples{
##--- load a dataset for demonstration
data("dt4rtn", package = "RTN")
gexp <- dt4rtn$gexp
annot <- dt4rtn$gexpIDs
tfs <- dt4rtn$tfs[c("IRF8","IRF1","PRDM1","AFF3","E2F3")]

##--- run mbrPreprocess
rmbr <- mbrPreprocess(gexp=gexp, regulatoryElements = tfs, 
rowAnnotation=annot)

##--- get the 'TNI' slot using 'mbrGet'
tni <- mbrGet(rmbr, what="TNI")

}
