% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllPlotsMBR.R
\name{mbrPlotDuals}
\alias{mbrPlotDuals}
\title{Plot shared targets between regulons.}
\usage{
mbrPlotDuals(object, dualreg, filepath = NULL, cols = c("#006400FF",
  "#CD6600FF"))
}
\arguments{
\item{object}{A processed object of class \linkS4class{MBR} evaluated by 
the method \code{\link{mbrAssociation}}.}

\item{dualreg}{A string indicating the name of a dual regulon.}

\item{filepath}{A string indicating the file path where the plot 
should be saved.}

\item{cols}{A vector of length 2 indicating a diverging color scheme 
for negative and positive correlations, respectively.}
}
\value{
A plot showing targets of dual regulons.
}
\description{
This function plots the shared targets for a regulon pair.
}
\examples{
##--- load a dataset for demonstration
data("dt4rtn", package = "RTN")
gexp <- dt4rtn$gexp
annot <- dt4rtn$gexpIDs
tfs <- dt4rtn$tfs[c("IRF8","IRF1","PRDM1","AFF3","E2F3")]

##--- run mbrPreprocess
rmbr <- mbrPreprocess(gexp=gexp, regulatoryElements = tfs, 
rowAnnotation=annot)

\dontrun{

##--- run mbrPermutation
rmbr <- mbrPermutation(rmbr, nPermutations=1000)

##--- run mbrBootstrap
rmbr <- mbrBootstrap(rmbr, nBootstrap=100)

##-- run mbrDpiFilter
rmbr <- mbrDpiFilter(rmbr)

##--- run mbrAssociation
rmbr <- mbrAssociation(rmbr, nPermutations=1000)

##--- get inferred duals and plot the shared cloud of targets
duals <- mbrGet(rmbr, what="dualRegulons")
mbrPlotDuals(rmbr, dualreg=duals[1])

}
}
