% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBR-methods.R
\docType{methods}
\name{tni2mbrPreprocess,TNI-method}
\alias{tni2mbrPreprocess,TNI-method}
\alias{tni2mbrPreprocess}
\title{A preprocessing function for objects of class MBR.}
\usage{
\S4method{tni2mbrPreprocess}{TNI}(tni, regulatoryElements = NULL,
  verbose = TRUE)
}
\arguments{
\item{tni}{A 'TNI' class object.}

\item{regulatoryElements}{A character vector specifying which 
'TNI' regulatory elements should be evaluated.}

\item{verbose}{A single logical value specifying to display detailed messages 
(when verbose=TRUE) or not (when verbose=FALSE).}
}
\value{
An \linkS4class{MBR} object.
}
\description{
This function converts a TNI class objects and into one MBR class object.
}
\examples{
#--- load a dataset for demonstration
data("dt4rtn", package = "RTN")
gexp <- dt4rtn$gexp
annot <- dt4rtn$gexpIDs
tfs <- dt4rtn$tfs[c("IRF8","IRF1","PRDM1","AFF3","E2F3")]

\dontrun{

##--- compute a TNI for tfs
tni <- tni.constructor(expData=gexp, regulatoryElements=tfs, rowAnnotation=annot)

##--- run tni2mbrPreprocess
rmbr <- tni2mbrPreprocess(tni)
}

}
