% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBR-methods.R
\docType{methods}
\name{mbrAssociation,MBR-method}
\alias{mbrAssociation}
\alias{mbrAssociation,MBR-method}
\title{Motifs analysis and inference of 'dual regulons'.}
\usage{
\S4method{mbrAssociation}{MBR}(object, regulatoryElements1 = NULL,
  regulatoryElements2 = NULL, minRegulonSize = 30, prob = 0.95,
  estimator = "spearman", pAdjustMethod = "BH", verbose = TRUE)
}
\arguments{
\item{object}{A processed object of class \linkS4class{MBR} evaluated by the 
methods \code{\link[RTNduals:mbrPermutation]{mbrPermutation}}, 
\code{\link[RTNduals:mbrBootstrap]{mbrBootstrap}} 
and \code{\link[RTNduals:mbrDpiFilter]{mbrDpiFilter}}.}

\item{regulatoryElements1}{An optional character vector specifying which 
'TNI1' regulatory elements should be evaluated. If 'NULL' all regulatory 
elements will be evaluated.}

\item{regulatoryElements2}{An optional character vector specifying which 
'TNI2' regulatory elements should be evaluated. If 'NULL' all regulatory 
elements will be evaluated.}

\item{minRegulonSize}{A single integer or numeric value specifying the 
minimum number of elements in a regulon. Gene sets with fewer than this 
number are removed from the analysis.}

\item{prob}{A quantile filter applyed to the association metric used to 
infer 'dual regulons'.}

\item{estimator}{A character value specifying the estimator used in the 
association analysis. One of "spearman" (default), "kendall", or "pearson", 
can be abbreviated.}

\item{pAdjustMethod}{A single character value specifying the p-value 
adjustment method to be used (see 'p.adjust' for details).}

\item{verbose}{A single logical value specifying to display detailed 
messages (when verbose=TRUE) or not (when verbose=FALSE).}
}
\value{
An \linkS4class{MBR} object with two data.frames in the slot 
'results' listing the inferred 'dual regulons' and a hypergeometric test 
for each 'dual regulon'.
}
\description{
This function takes an MBR object and compares the shared regulon 
targets in order to test whether regulon pairs agree on the predicted 
downstream effects.
}
\examples{
data("dt4rtn", package = "RTN")
gexp <- dt4rtn$gexp
annot <- dt4rtn$gexpIDs
tfs1 <- dt4rtn$tfs[c("IRF8","IRF1","PRDM1","AFF3","E2F3")]
tfs2 <- dt4rtn$tfs[c("HCLS1","STAT4","STAT1","LMO4","ZNF552")]
##---mbrPreprocess
rmbr <- mbrPreprocess(gexp=gexp, regulatoryElements1 = tfs1, 
regulatoryElements2=tfs2, gexpIDs=annot)
##---mbrPermutation
rmbr <- mbrPermutation(rmbr, nPermutations=10)
##---mbrBootstrap
rmbr <- mbrBootstrap(rmbr, nBootstrap=10)
##---mbrDpiFilter
rmbr <- mbrDpiFilter(rmbr)
##---mbrAssociation
rmbr <- mbrAssociation(rmbr, prob=0.75)

}

