% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBR-methods.R
\docType{methods}
\name{mbrBootstrap,MBR-method}
\alias{mbrBootstrap}
\alias{mbrBootstrap,MBR-method}
\title{Inference of consensus transcriptional networks.}
\usage{
\S4method{mbrBootstrap}{MBR}(object, verbose = TRUE, ...)
}
\arguments{
\item{object}{A processed objec of class \linkS4class{MBR} evaluated by the 
method \code{\link[RTNduals:mbrPermutation]{mbrPermutation}}.}

\item{verbose}{A single logical value specifying to display detailed 
messages (when verbose=TRUE) or not (when verbose=FALSE).}

\item{...}{Additional arguments passed to the 
\code{\link[RTN:tni.bootstrap]{tni.bootstrap}} function.}
}
\value{
An \linkS4class{MBR} object with two consensus mutual information 
matrices, one in each "TNI" slot.
}
\description{
This function takes an MBR object and computes two consensus transcriptional 
networks.
}
\examples{
data("dt4rtn", package = "RTN")
gexp <- dt4rtn$gexp
annot <- dt4rtn$gexpIDs
tfs1 <- dt4rtn$tfs[c("IRF8","IRF1","PRDM1","AFF3","E2F3")]
tfs2 <- dt4rtn$tfs[c("HCLS1","STAT4","STAT1","LMO4","ZNF552")]
##---mbrPreprocess
rmbr <- mbrPreprocess(gexp=gexp, regulatoryElements1 = tfs1, 
regulatoryElements2=tfs2, gexpIDs=annot)
##---mbrPermutation
rmbr <- mbrPermutation(rmbr, nPermutations=10)
##---mbrBootstrap
rmbr <- mbrBootstrap(rmbr, nBootstrap=10)

}

