% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllPlotsMBR.R
\name{mbrPlotDuals}
\alias{mbrPlotDuals}
\title{Plot shared target clouds between dual regulons.}
\usage{
mbrPlotDuals(object, names.motifs = NULL, filepath = NULL, alpha = 0.8,
  lncols = c("darkgreen", "darkorange3"), lwd = 0.7,
  estimator = "spearman")
}
\arguments{
\item{object}{A processed object of class \linkS4class{MBR} evaluated by 
the method \code{\link[RTNduals:mbrAssociation]{mbrAssociation}}.}

\item{names.motifs}{A vector with 'dual regulon' indentifiers from the 
'motifsInformation' table.}

\item{filepath}{A character string indicating the file path where the plot 
should be saved.}

\item{alpha}{The alpha transparency, a number in [0,1].}

\item{lncols}{A vector of length 2 indicating the colors of the negative 
and positive target clouds, respectively.}

\item{lwd}{Line width, a decimal value (between 0 and 1).}

\item{estimator}{A character string indicating the association metric. 
One of "spearman" (default), "kendall", or "pearson", can be abbreviated.}
}
\value{
A plot with the shared target clouds between dual regulons.
}
\description{
This function plots the shared target clouds between a regulon pair.
}
\examples{
data("dt4rtn", package = "RTN")
gexp <- dt4rtn$gexp
annot <- dt4rtn$gexpIDs
tfs1 <- dt4rtn$tfs[c("IRF8","IRF1","PRDM1","AFF3","E2F3")]
tfs2 <- dt4rtn$tfs[c("HCLS1","STAT4","STAT1","LMO4","ZNF552")]
##---mbrPreprocess
rmbr <- mbrPreprocess(gexp=gexp, regulatoryElements1 = tfs1, 
regulatoryElements2=tfs2, gexpIDs=annot)
##---mbrPermutation
rmbr <- mbrPermutation(rmbr, nPermutations=10)
##---mbrBootstrap
rmbr <- mbrBootstrap(rmbr, nBootstrap=10)
##---mbrDpiFilter
rmbr <- mbrDpiFilter(rmbr)
##---mbrAssociation
rmbr <- mbrAssociation(rmbr, prob=0.75)
##---mbrDuals
rmbr <- mbrDuals(rmbr)
##---
dual <- mbrGet(rmbr, what="dualRegulons")[1]
mbrPlotDuals(rmbr, names.motifs=dual)

}

