% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBR-methods.R
\docType{methods}
\name{tni2mbrPreprocess,TNI-method}
\alias{tni2mbrPreprocess}
\alias{tni2mbrPreprocess,TNI-method}
\title{A preprocessing function for objects of class MBR.}
\usage{
\S4method{tni2mbrPreprocess}{TNI}(tni1, tni2, verbose = TRUE)
}
\arguments{
\item{tni1}{A 'TNI' class object.}

\item{tni2}{Another 'TNI' class object}

\item{verbose}{A single logical value specifying to display detailed messages 
(when verbose=TRUE) or not (when verbose=FALSE).}
}
\value{
An \linkS4class{MBR} object.
}
\description{
This function merges two TNI class objects and creates one MBR class object.
}
\examples{
data("dt4rtn", package = "RTN")
gexp <- dt4rtn$gexp
annot <- dt4rtn$gexpIDs
tfs1 <- dt4rtn$tfs[c("IRF8","IRF1","PRDM1","AFF3","E2F3")]
tfs2 <- dt4rtn$tfs[c("HCLS1","STAT4","STAT1","LMO4","ZNF552")]
##---mbrPreprocess
rmbr <- mbrPreprocess(gexp=gexp, regulatoryElements1 = tfs1, 
regulatoryElements2=tfs2, gexpIDs=annot)
##---mbrPermutation
rmbr <- mbrPermutation(rmbr, nPermutations=10)
##---mbrBootstrap
rmbr <- mbrBootstrap(rmbr, nBootstrap=10)
rmbr <- mbrDpiFilter(rmbr)
##---tni2mbrPreprocess
tni1 <- mbrGet(rmbr, what="TNI1")
tni2 <- mbrGet(rmbr, what="TNI2")
rmbr <- tni2mbrPreprocess(tni1, tni2)

}

